Extract devtools renderables into own module
